changed everything to absolute imports, and added feature to convert binary stl stuff to 64 bit values for easier computation

+++ utils/config.go
@@ -6,7 +6,7 @@ import (
 )
 
 type Config struct {
-  LayerHeight float32
+  LayerHeight float64
   HeatBed bool
   BedTemp int
   ExtruderTemp int
